Fix missing hulls when concave polygons are not generated for groups
Polygons are not run through `concaveman` for certain conditions, such as polygons with 2 or fewer points. The original matrices have `xx` and `yy` column names, which are not converted to the names generated from `concaveman` and are not displayed. Fix by renaming all the matrices' columns to the same set of names.